Move App to the components directory since it is a presentation component
